+++ Src/Notion.Client/Models/PropertyValue/DateJsonObject.cs
@@ -1,0 +1,16 @@
+using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    internal class DateJsonObject
+    {
+        [JsonProperty("start")]
+        public string Start { get; set; }
+
+        [JsonProperty("end")]
+        public string End { get; set; }
+
+        [JsonProperty("time_zone")]
+        public string TimeZone { get; set; }
+    }
+}
